moved all calendar specific helpers to be under the calendar utils
