+++ frontend/src/functionality/Calendar/utils/Api/savePlan.js
@@ -1,4 +1,4 @@
-import { api } from '../../api.js';
+import { api } from '../../../../api.js';
 
 /**
  * Saves a generated schedule
