+++ frontend/src/functionality/Calendar/utils/Helpers/discardSchedule.js
